requirements.txt
pandas
sklearn
numpy
scikit-surprise